Fix poetry-get-virtualenv
This was previously assuming the name of the poetry virtualenv; however,
it was incorrect (as the latest version of poetry appends more to the
name, e.g. `projectname-Q5nLX01s-py3.7`). Instead of guessing, we can
call `poetry env list`.

This can be further improved, as `poetry env list --full-path` does as
it suggest, and provides the full path which we can use instead of
creating it ourselves.

This may introduce a problem because I think this can potentially list
multiple virtualenvs associated with the project, and we'd need to
figure out which is the current one in use (or ask the user to choose,
perhaps).