Fix toml parsing
According to the TOML specs in https://toml.io/en/v1.0.0, A hash symbol
marks the rest of the line as a comment, except when inside a string.

Currently, poetry.el fails to detect a project when the toml file is:

```
...
[tool.poetry] # some comment
...
```

So running `poetry-show` ends with a:
```
poetry-error: [Poetry] Not in a poetry project
```

After removing the end of line matching of the regex, the project is
successfully loaded.